refactor(folder structure) : make buffer folder
move Buffer class & concrete class files to buffer folder
