+++ src/Webserv.hpp
@@ -16,11 +16,11 @@
 #include "server/Server.hpp"
 #include "parser/ConfigParsing.hpp"
 #include "server/ServerConfig.hpp"
-#include "Buffer.hpp"
-#include "Message.hpp"
-#include "File.hpp"
+#include "buffer/Buffer.hpp"
+#include "buffer/Message.hpp"
+#include "buffer/File.hpp"
 #include "parser/RequestParser.hpp"
-#include "Library.hpp"
+#include "buffer/Library.hpp"
 
 class Server;
