+++ src/parser/CgiParser.hpp
@@ -15,7 +15,7 @@
 
 #include <cctype>
 #include <stdexcept>
-#include "../Library.hpp"
+#include "../buffer/Library.hpp"
 
 class CgiParser{
 	public:
